Extract ProjectController
